Add retrieve comments functionality

+++ SUSSYCourses/src/main/java/com/sustech/cs309/project/sussycourses/repository/CoursewareRepository.java
@@ -13,7 +13,7 @@ public interface CoursewareRepository extends JpaRepository<Course, Long> {
     Optional<Course> findById(Long id);
 
     @Query("SELECT cw FROM Courseware cw JOIN cw.course c WHERE c.courseId = :courseId")
-    List<Courseware> findCoursewareForCourse(@Param("courseId") int courseId);
+    List<Courseware> findCoursewareForCourse(@Param("courseId") long courseId);
 
 
 }
